move all reducers to reducers folder
