+++ src/reducers/sidebarReducer.js
